+++ lib/bumblebee/multimodal/layout_lm.ex
@@ -177,11 +177,11 @@ defmodule Bumblebee.Multimodal.LayoutLm do
 
   @impl true
   def input_template(%{architecture: :for_multiple_choice}) do
-    %{"input_ids" => Nx.template({1, 1, 1}, :s64)}
+    %{"input_ids" => Nx.template({1, 1, 1}, :u32)}
   end
 
   def input_template(_spec) do
-    %{"input_ids" => Nx.template({1, 1}, :s64)}
+    %{"input_ids" => Nx.template({1, 1}, :u32)}
   end
 
   @impl true
