+++ lib/bumblebee/text/bart.ex
@@ -232,7 +232,7 @@ defmodule Bumblebee.Text.Bart do
   @impl true
   def input_template(_spec) do
     %{
-      "input_ids" => Nx.template({1, 1}, :s64)
+      "input_ids" => Nx.template({1, 1}, :u32)
     }
   end
