+++ lib/bumblebee/text/clip_text.ex
@@ -107,7 +107,7 @@ defmodule Bumblebee.Text.ClipText do
   @impl true
   def input_template(_spec) do
     %{
-      "input_ids" => Nx.template({1, 1}, :s64)
+      "input_ids" => Nx.template({1, 1}, :u32)
     }
   end
