+++ lib/bumblebee/text/gpt2.ex
@@ -182,7 +182,7 @@ defmodule Bumblebee.Text.Gpt2 do
 
   @impl true
   def input_template(_spec) do
-    %{"input_ids" => Nx.template({1, 1}, :s64)}
+    %{"input_ids" => Nx.template({1, 1}, :u32)}
   end
 
   @impl true
