+++ lib/bumblebee/text/zero_shot_classification.ex
@@ -57,9 +57,9 @@ defmodule Bumblebee.Text.ZeroShotClassification do
           Shared.compile_or_jit(scores_fun, defn_options, compile != nil, fn ->
             inputs = %{
               "input_ids" =>
-                Nx.template({batch_size, sequences_per_batch, sequence_length}, :s64),
+                Nx.template({batch_size, sequences_per_batch, sequence_length}, :u32),
               "attention_mask" =>
-                Nx.template({batch_size, sequences_per_batch, sequence_length}, :s64)
+                Nx.template({batch_size, sequences_per_batch, sequence_length}, :u32)
             }
 
             [params, inputs]
